Improve argument name checks for build/compute_output_signature
We want to support the case where a layer looks like this:

```python
class SomeLayer(keras_core.layers.Layer):
    def build(self, shape):
        ...

    def call(
        self,
        inputs,
        optional_tensor_arg_one=None,
        optional_tensor_arg_two=None,
        optional_tensor_arg_tree=None,
    ):
        ...
```

However, due to a bug, we would forward a dictionary of all passed
tensor shapes to build when it only accepted one argument.

This tries to clean up the logic a bit.