Copy torch tensor before update in `slice_update`
Without this change, the `slice_update` call would fail when gradients
are need for the input/output tensors:

```
RuntimeError: a leaf Variable that requires grad is being used in an in-place operation.
```

With this change, the error goes away as the slice is no longer
considered in place for the input variable.