Fixed SeedGenerator in the torch backend
We were ignoring the second seed entirely, which is quite incorrect as
that is the only part of the seed we are incrementing.

I'm just summing the fixed seed and increment counter together for now.
Not sure if that's the best approach for truly random seeding, or
what approach torch is even using for there seeds, but figure this is
much better than nothing.