Update dtype argument for keras_core.Input
At present,the API keras_core.input documentation states that the argument dtype to be passed as a string such as 'tf.float32' etc. However if we pass `tf.dtypes.DType` such as `tf.float32`,`tf.float64` or  a `torch.dtype` such as `torch.float32`,`torch.float64` it will works as well as the internal code calls keras_core.backend.standardize_dtype which can convert `tf.float32` or `torch.float32` to string such as 'float32'.

IMO, its better to mention this is documentation that this API can also handle both `tf.dtypes.DType` and `torch.dtype` so that users are aware of this and they are free to choose any of the above ways to pass value to dtype.